LayoutTests/webrtc/peer-connection-remote-audio-mute2.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Testing local audio capture playback causes "playing" event to fire</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
</head>
<body>
    <script src ="routines.js"></script>
    <script>
    promise_test((test) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);

        return navigator.mediaDevices.getUserMedia({audio: true}).then((localStream) => {
            if (window.internals)
                internals.useMockRTCPeerConnectionFactory("TwoRealPeerConnections");

            var remoteTrack;
            var remoteStream;
            return new Promise((resolve, reject) => {
                createConnections((firstConnection) => {
                    firstConnection.addStream(localStream);
                }, (secondConnection) => {
                    secondConnection.onaddstream = (streamEvent) => {
                        remoteStream = streamEvent.stream;
                        remoteTrack = remoteStream.getAudioTracks()[0];
                        resolve();
                    };
                });
            }).then(() => {
                return analyseAudio(remoteStream, 500).then((results) => {
                    assert_true(results.heardHum, "heard hum from remote enabled track");
                });
            }).then(() => {
                remoteTrack.enabled = false;
            }).then(() => {
                return analyseAudio(remoteStream, 500).then((results) => {
                    assert_false(results.heardHum, "not heard hum from remote disabled track");
                });
            }).then(() => {
                remoteTrack.enabled = true;
            }).then(() => {
                return analyseAudio(remoteStream, 500).then((results) => {
                    assert_true(results.heardHum, "heard hum from remote reenabled track");
                });
            });
        });
    }, "Muting and unmuting an incoming audio track");
    </script>
</body>
</html>
